Small restructuring
