+++ src/errors.rs
@@ -1,7 +1,6 @@
 #[allow(unused)]
 use crate::prelude::*;
 
-use crate::Value;
 use derive_new::new;
 
 #[derive(Debug, Ord, PartialOrd, Eq, PartialEq, new)]
