+++ src/format/generic.rs
@@ -1,6 +1,6 @@
 use crate::format::{EntriesView, RenderView, TableView};
 use crate::object::Value;
-use crate::Host;
+use crate::prelude::*;
 use derive_new::new;
 
 // A list is printed one line at a time with an optional separator between groups
