+++ src/format/list.rs
@@ -1,5 +1,5 @@
 use crate::format::RenderView;
-use crate::Host;
+use crate::prelude::*;
 use derive_new::new;
 
 // A list is printed one line at a time with an optional separator between groups
