+++ src/format/table.rs
@@ -1,6 +1,6 @@
 use crate::format::RenderView;
 use crate::object::Value;
-use crate::Host;
+use crate::prelude::*;
 use derive_new::new;
 use prettytable::{Cell, Row, Table};
