+++ src/prelude.rs
@@ -1,4 +1,6 @@
-crate use crate::commands::command::{Command, CommandArgs, ReturnValue};
+crate use crate::cli::MaybeOwned;
+crate use crate::commands::command::{Command, CommandAction, CommandArgs, ReturnValue};
+crate use crate::context::Context;
 crate use crate::env::{Environment, Host};
 crate use crate::errors::ShellError;
 crate use crate::object::{Primitive, Value};
